feat(mobile): extract PageList component
