+++ apps/mobile/src/modules/entry-list/index.tsx
@@ -1,10 +1,9 @@
 import { FeedViewType } from "@follow/constants"
-import * as Haptics from "expo-haptics"
-import { useEffect, useMemo } from "react"
-import { Animated, StyleSheet } from "react-native"
-import PagerView from "react-native-pager-view"
+import { useMemo } from "react"
 
-import { selectTimeline, useSelectedFeed, useSelectedView } from "@/src/modules/screen/atoms"
+import { useSelectedFeed, useSelectedView } from "@/src/modules/screen/atoms"
+import { PagerList } from "@/src/modules/screen/PageList"
+import { TimelineSelectorProvider } from "@/src/modules/screen/TimelineSelectorProvider"
 import {
   useEntryIdsByCategory,
   useEntryIdsByFeedId,
@@ -12,12 +11,8 @@ import {
   useEntryIdsByListId,
   useEntryIdsByView,
 } from "@/src/store/entry/hooks"
-import { useViewWithSubscription } from "@/src/store/subscription/hooks"
 
-import { TimelineSelectorProvider } from "../screen/TimelineSelectorProvider"
-import { setHorizontalScrolling } from "./atoms"
 import { EntryListSelector } from "./EntryListSelector"
-import { usePagerView } from "./usePagerView"
 
 export function EntryList() {
   const selectedFeed = useSelectedFeed()
@@ -26,7 +21,13 @@ export function EntryList() {
     if (!selectedFeed) return null
     switch (selectedFeed.type) {
       case "view": {
-        return <ViewPagerList viewId={selectedFeed.viewId} />
+        return (
+          <PagerList
+            renderItem={(view, active) => (
+              <ViewEntryList key={view} viewId={view} active={active} />
+            )}
+          />
+        )
       }
       case "feed": {
         return <FeedEntryList feedId={selectedFeed.feedId} />
@@ -45,54 +46,6 @@ export function EntryList() {
   if (!Content) return null
 
   return <TimelineSelectorProvider>{Content}</TimelineSelectorProvider>
-}
-
-const AnimatedPagerView = Animated.createAnimatedComponent<typeof PagerView>(PagerView)
-
-function ViewPagerList({ viewId }: { viewId: FeedViewType }) {
-  const activeViews = useViewWithSubscription()
-  const viewIdIndex = activeViews.findIndex((view) => view.view === viewId)
-  const { page, pagerRef, ...rest } = usePagerView({
-    initialPage: viewIdIndex,
-    onIndexChange: (index) => {
-      selectTimeline({ type: "view", viewId: activeViews[index]!.view })
-    },
-  })
-
-  useEffect(() => {
-    if (page === viewIdIndex) return
-    pagerRef.current?.setPage(viewIdIndex)
-  }, [page, pagerRef, viewIdIndex])
-
-  return (
-    <AnimatedPagerView
-      testID="pager-view"
-      ref={pagerRef}
-      style={styles.PagerView}
-      initialPage={page}
-      layoutDirection="ltr"
-      overdrag
-      onPageScroll={rest.onPageScroll}
-      onPageSelected={rest.onPageSelected}
-      onPageScrollStateChanged={(e) => {
-        rest.onPageScrollStateChanged(e)
-        setHorizontalScrolling(e.nativeEvent.pageScrollState !== "idle")
-        if (e.nativeEvent.pageScrollState === "settling") {
-          Haptics.impactAsync(Haptics.ImpactFeedbackStyle.Light)
-        }
-      }}
-      pageMargin={10}
-      orientation="horizontal"
-    >
-      {useMemo(
-        () =>
-          activeViews.map((view, index) => (
-            <ViewEntryList key={view.view} viewId={view.view} active={page === index} />
-          )),
-        [activeViews, page],
-      )}
-    </AnimatedPagerView>
-  )
 }
 
 function ViewEntryList({ viewId, active }: { viewId: FeedViewType; active: boolean }) {
@@ -124,12 +77,3 @@ function InboxEntryList({ inboxId }: { inboxId: string }) {
   const entryIds = useEntryIdsByInboxId(inboxId)
   return <EntryListSelector entryIds={entryIds} viewId={view} />
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  PagerView: {
-    flex: 1,
-  },
-})
