+++ apps/mobile/src/modules/screen/atoms.ts
@@ -224,3 +224,12 @@ export const useViewDefinition = (view?: FeedViewType) => {
   const viewDef = useMemo(() => views.find((v) => v.view === view), [view])
   return viewDef
 }
+
+// horizontal scrolling state
+
+const horizontalScrollingAtom = atom<boolean>(false)
+
+export const setHorizontalScrolling = (value: boolean) =>
+  jotaiStore.set(horizontalScrollingAtom, value)
+
+export const getHorizontalScrolling = () => jotaiStore.get(horizontalScrollingAtom)
